Load partial solutions
Whenever a collection is loaded, try to load not only the first level the
player has not yet solved, but the list of moves the player made prior to
saving. In addition loading a modified level, the list of moves is
preserved, so undo/redo works *exactly* like before saving and exiting.