Simplify input handling
Move recording_macro and cursor_pos to a separate structure, make an
instance of this and command_queue local variables of main_loop().

Use the modifiers field of keyboard events so we do not have to keep track
of which modifier key is currently pressed by handling Pressed and Released
events.