Don't redraw the window when it is out of focus
This is the first step in creating the infrastructure for only calling
Game::render if the screen really needs to be updated. Not updating
out-of-focus windows is just a very simple heuristic. It does not handle
animations correctly, i.e. when an animation is running and the window
loses focus, the last animation step rendered will remain visible until
the window regains focus.