Remove Game::test::switch_levels()
This test relies on the filesystem being in a particular state.
Specifically, it needs to be run in a world where level 1 of the
collection test has been solved.

Until the code loading this information from disk can be replaced by a
dummy, the test cannot run correctly on Travis.